+++ packages/neuron-ui/.eslintrc.js
@@ -11,5 +11,6 @@ module.exports = {
     "react/jsx-filename-extension": [1, {
       "extensions": [".ts", ".tsx"]
     }],
+    "react/jsx-props-no-spreading": [0]
   }
 }