+++ packages/neuron-ui/src/states/stateProvider/actionCreators/transactions.ts
@@ -1,4 +1,4 @@
-import { NeuronWalletActions, AppActions, StateDispatch } from 'states/stateProvider/reducer'
+import { NeuronWalletActions, StateDispatch } from 'states/stateProvider/reducer'
 import {
   GetTransactionListParams,
   getTransactionList,
@@ -23,34 +23,24 @@ export const updateTransactionList = (params: GetTransactionListParams) => (disp
 export const updateTransactionDescription = (params: Controller.UpdateTransactionDescriptionParams) => (
   dispatch: StateDispatch
 ) => {
+  const descriptionParams = {
+    hash: params.hash,
+    description: params.description,
+  }
   dispatch({
-    type: AppActions.UpdateLoadings,
-    payload: {
-      updateDescription: true,
-    },
-  })
-  updateRemoteTransactionDescription(params)
-    .then(res => {
-      if (res.status) {
-        dispatch({
-          type: NeuronWalletActions.UpdateTransactionDescription,
-          payload: {
-            hash: params.hash,
-            description: params.description,
-          },
-        })
-      } else {
-        addNotification(failureResToNotification(res))(dispatch)
-      }
-    })
-    .finally(() => {
+    type: NeuronWalletActions.UpdateTransactionDescription,
+    payload: descriptionParams,
+  }) // update local description before remote description to avoid the flicker on the field
+  updateRemoteTransactionDescription(params).then(res => {
+    if (res.status) {
       dispatch({
-        type: AppActions.UpdateLoadings,
-        payload: {
-          updateDescription: false,
-        },
+        type: NeuronWalletActions.UpdateTransactionDescription,
+        payload: descriptionParams,
       })
-    })
+    } else {
+      addNotification(failureResToNotification(res))(dispatch)
+    }
+  })
 }
 
 export default {