+++ packages/neuron-wallet/src/decorators/errors.ts
@@ -1,4 +1,5 @@
 import { ResponseCode } from 'utils/const'
+import logger from 'utils/logger'
 
 export const CatchControllerError = (_target: any, _name: string, descriptor: PropertyDescriptor) => {
   const originalMethod = descriptor.value
@@ -8,6 +9,7 @@ export const CatchControllerError = (_target: any, _name: string, descriptor: Pr
       try {
         return await originalMethod(...args)
       } catch (err) {
+        logger.error(`CatchControllerError:`, err)
         return {
           status: ResponseCode.Fail,
           message: typeof err.message === 'string' ? { content: err.message } : err.message,