+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+# [0.19.0-beta.1](https://github.com/nervosnetwork/neuron/compare/v0.19.0-beta.0...v0.19.0-beta.1) (2019-09-04)
+
+
+### Bug Fixes
+
+* add totalBalance to address test ([90184aa](https://github.com/nervosnetwork/neuron/commit/90184aa))
+* change typeScript type to `text` from `varchar` ([e334085](https://github.com/nervosnetwork/neuron/commit/e334085))
+* fix balance tests ([11dcac7](https://github.com/nervosnetwork/neuron/commit/11dcac7))
+* network test import ([c4cb7e4](https://github.com/nervosnetwork/neuron/commit/c4cb7e4))
+
+
+### Features
+
+* add base settings and move skip config to here ([37d82fb](https://github.com/nervosnetwork/neuron/commit/37d82fb))
+* **neuron-ui:** move the miner info from overview to wallet list ([#907](https://github.com/nervosnetwork/neuron/issues/907)) ([1378768](https://github.com/nervosnetwork/neuron/commit/1378768))
+* add an edit button by the side of description ([031fed0](https://github.com/nervosnetwork/neuron/commit/031fed0))
+* add controller for skip data and type ([834390f](https://github.com/nervosnetwork/neuron/commit/834390f))
+* add hasData and typeScript to output entity ([9fe535b](https://github.com/nervosnetwork/neuron/commit/9fe535b))
+* add SkipDataAndType class ([a952152](https://github.com/nervosnetwork/neuron/commit/a952152))
+* add totalBalance to address entity ([5746438](https://github.com/nervosnetwork/neuron/commit/5746438))
+* add totalBalance to address interface ([7a29c03](https://github.com/nervosnetwork/neuron/commit/7a29c03))
+* calculate totalBalance and check skip data and type in gather inputs ([f455545](https://github.com/nervosnetwork/neuron/commit/f455545))
+* **neuron-ui:** prevent updating tx and addr list when user is editing the description ([6c4ea72](https://github.com/nervosnetwork/neuron/commit/6c4ea72))
+* skip cells which has data or type ([6a85705](https://github.com/nervosnetwork/neuron/commit/6a85705))
+* using keep alive connection ([f3fbe15](https://github.com/nervosnetwork/neuron/commit/f3fbe15))
+
+
+
 # [0.19.0-beta.0](https://github.com/nervosnetwork/neuron/compare/v0.18.0-beta.1...v0.19.0-beta.0) (2019-08-29)
 
 ### BREAKING CHANGES